Implement expense and income delete end point

+++ backend/src/main/java/com/pm/moneymanager/repository/ExpenseRepository.java
@@ -11,7 +11,7 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface ExpenseRepository extends JpaRepository<Expense, Integer> {
+public interface ExpenseRepository extends JpaRepository<Expense, Long> {
 
     List<Expense> findByProfileIdOrderByDateDesc(Long profileId);
